refactor: change the name of commons componente to put in ui folders
